make size_t

+++ src/convolution_api.cpp
@@ -232,9 +232,9 @@ miopenConvolutionCKBackwardWeightsGetWorkSpaceSize(const miopenAlphaBetaCase_t a
                                          size_t K,
                                          miopenDataType_t data_type,
                                          miopenAlphaBetaCase_t alpha_beta_case) {
-            auto is_odd     = [](int num) { return num % 2 != 0; };
-            int C_per_group = C / G;
-            int K_per_group = K / G;
+            auto is_odd        = [](int num) { return num % 2 != 0; };
+            size_t C_per_group = C / G;
+            size_t K_per_group = K / G;
 
             return (alpha_beta_case == BILINEAR || alpha_beta_case == SCALE) ||
                    (data_type == miopenHalf &&
